+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/EditorConfig.kt
@@ -2,6 +2,7 @@ package com.pinterest.ktlint.core
 
 import com.pinterest.ktlint.core.EditorConfig.IndentStyle.SPACE
 import com.pinterest.ktlint.core.EditorConfig.IndentStyle.TAB
+import com.pinterest.ktlint.core.api.DefaultEditorConfigProperties
 import com.pinterest.ktlint.core.api.FeatureInAlphaState
 import com.pinterest.ktlint.core.api.UsesEditorConfigProperties
 import org.ec4j.core.model.PropertyType
@@ -12,29 +13,46 @@ import org.jetbrains.kotlin.com.intellij.lang.ASTNode
  *
  * This class is injected into the user data, so it is available to rules via [KtLint.EDITOR_CONFIG_USER_DATA_KEY]
  */
+@Deprecated(
+    message = "Marked for removal in Ktlint 0.46. Implement interface UsesEditorConfigProperties on the rule and " +
+        "retrieve values via call 'node.getEditorConfigValue(property)'."
+)
 interface EditorConfig {
-    @Deprecated("Replace with IndentConfig.IndentStyle")
+    @Deprecated("Marked for removal in Ktlint 0.46. Replace with IndentConfig.IndentStyle")
     enum class IndentStyle { SPACE, TAB }
 
-    @Deprecated("Prefer to use IndentConfig.indent only or IndentConfig.indentStyle otherwise")
+    @Deprecated("Marked for removal in Ktlint 0.46. Use IndentConfig.indent only or IndentConfig.indentStyle otherwise")
     public val indentStyle: IndentStyle
 
-    @Deprecated("Prefer to use IndentConfig.indent.length")
+    @Deprecated("Marked for removal in Ktlint 0.46. Prefer to use IndentConfig.indent.length")
     val indentSize: Int
 
-    @Deprecated("Prefer to use IndentConfig.indent.length")
+    @Deprecated("Marked for removal in Ktlint 0.46. Prefer to use IndentConfig.indent.length")
     val tabWidth: Int
 
+    @Deprecated(
+        message = "Marked for removal in Ktlint 0.46. Implement interface UsesEditorConfigProperties on the rule and " +
+            "retrieve this value via call 'node.getEditorConfigValue(maxLineLengthProperty)'."
+    )
     val maxLineLength: Int
 
     @Deprecated(
-        message = "Not used anymore by rules, please use 'insert_final_newline' directly via get()"
+        message = "Marked for removal in Ktlint 0.46. Implement interface UsesEditorConfigProperties on the rule and " +
+            "retrieve this value via call 'node.getEditorConfigValue(insertNewLineProperty)'."
     )
     val insertFinalNewline: Boolean
 
+    @Deprecated(
+        message = "Marked for removal in Ktlint 0.46. Implement interface UsesEditorConfigProperties on the rule and " +
+            "retrieve values via call 'node.getEditorConfigValue(property)'."
+    )
     operator fun get(key: String): String?
 
     companion object {
+        @Deprecated(
+            message = "Marked for removal in Ktlint 0.46. Implement interface UsesEditorConfigProperties on the " +
+                "rule and retrieve values via call 'node.getEditorConfigValue(property)'."
+        )
         fun fromMap(map: Map<String, String>): EditorConfig {
             val indentStyle = when {
                 map["indent_style"]?.toLowerCase() == "tab" -> TAB
@@ -62,18 +80,29 @@ interface EditorConfig {
             }
         }
 
+        @Deprecated(
+            message = "Marked for removal in Ktlint 0.46. The interface UsesEditorConfigProperties needs to be " +
+                "implemented first on the rule.",
+            replaceWith = ReplaceWith("this.getEditorConfigValue(property)")
+        )
         public fun ASTNode.loadEditorConfig(): EditorConfig = getUserData(KtLint.EDITOR_CONFIG_USER_DATA_KEY)!!
 
+        @Deprecated(
+            message = "Marked for removal in Ktlint 0.46. Implement interface UsesEditorConfigProperties on the " +
+                "rule. Then configure the IndentConfig inside the rule.",
+            replaceWith = ReplaceWith(
+                "IndentConfig(indentStyle = IndentStyleValue, tabWidth = tabWidth)",
+                "com.pinterest.ktlint.core.IndentConfig"
+            )
+        )
         public fun EditorConfig.loadIndentConfig(): IndentConfig =
             IndentConfig(
-                indentStyle = when (indentStyle) {
-                    TAB -> IndentConfig.IndentStyle.TAB
-                    SPACE -> IndentConfig.IndentStyle.SPACE
-                },
+                indentStyle = indentStyle,
                 tabWidth = tabWidth,
                 disabled = indentSize <= 0 || tabWidth <= 0
             )
 
+        @Deprecated("Marked for removal in Ktlint 0.46.")
         /**
          * Use this value to define a non-nullable class variable of type EditorConfig. When loading the root node in
          * rule visitor, the value should be replaced with the real value.
@@ -96,18 +125,20 @@ interface EditorConfig {
             inner class EditorConfigNotInitializedException : RuntimeException("EditorConfig is not yet initialized")
         }
 
+        @Deprecated(
+            message = "Marked for removal in Ktlint 0.46",
+            replaceWith = ReplaceWith("DefaultEditorConfigProperties.indentStyleProperty")
+        )
         @OptIn(FeatureInAlphaState::class)
         public val indentStyleProperty: UsesEditorConfigProperties.EditorConfigProperty<PropertyType.IndentStyleValue> =
-            UsesEditorConfigProperties.EditorConfigProperty(
-                type = PropertyType.indent_style,
-                defaultValue = PropertyType.IndentStyleValue.space
-            )
+            DefaultEditorConfigProperties.indentStyleProperty
 
+        @Deprecated(
+            message = "Marked for removal in Ktlint 0.46",
+            replaceWith = ReplaceWith("DefaultEditorConfigProperties.indentSizeProperty")
+        )
         @OptIn(FeatureInAlphaState::class)
         public val indentSizeProperty: UsesEditorConfigProperties.EditorConfigProperty<Int> =
-            UsesEditorConfigProperties.EditorConfigProperty(
-                type = PropertyType.indent_size,
-                defaultValue = IndentConfig.DEFAULT_INDENT_CONFIG.tabWidth
-            )
+            DefaultEditorConfigProperties.indentSizeProperty
     }
 }