+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/IndentationRule.kt
@@ -1,13 +1,13 @@
 package com.pinterest.ktlint.ruleset.standard
 
-import com.pinterest.ktlint.core.EditorConfig
-import com.pinterest.ktlint.core.EditorConfig.Companion.loadEditorConfig
-import com.pinterest.ktlint.core.EditorConfig.Companion.loadIndentConfig
 import com.pinterest.ktlint.core.IndentConfig
 import com.pinterest.ktlint.core.IndentConfig.IndentStyle.SPACE
 import com.pinterest.ktlint.core.IndentConfig.IndentStyle.TAB
 import com.pinterest.ktlint.core.Rule
-import com.pinterest.ktlint.core.ast.ElementType.ANNOTATION
+import com.pinterest.ktlint.core.api.DefaultEditorConfigProperties.indentSizeProperty
+import com.pinterest.ktlint.core.api.DefaultEditorConfigProperties.indentStyleProperty
+import com.pinterest.ktlint.core.api.FeatureInAlphaState
+import com.pinterest.ktlint.core.api.UsesEditorConfigProperties
 import com.pinterest.ktlint.core.ast.ElementType.ARROW
 import com.pinterest.ktlint.core.ast.ElementType.BINARY_EXPRESSION
 import com.pinterest.ktlint.core.ast.ElementType.BINARY_WITH_TYPE
@@ -17,9 +17,9 @@ import com.pinterest.ktlint.core.ast.ElementType.BY_KEYWORD
 import com.pinterest.ktlint.core.ast.ElementType.CALL_EXPRESSION
 import com.pinterest.ktlint.core.ast.ElementType.CLOSING_QUOTE
 import com.pinterest.ktlint.core.ast.ElementType.COLON
-import com.pinterest.ktlint.core.ast.ElementType.COMMA
 import com.pinterest.ktlint.core.ast.ElementType.CONDITION
 import com.pinterest.ktlint.core.ast.ElementType.DELEGATED_SUPER_TYPE_ENTRY
+import com.pinterest.ktlint.core.ast.ElementType.DOT
 import com.pinterest.ktlint.core.ast.ElementType.DOT_QUALIFIED_EXPRESSION
 import com.pinterest.ktlint.core.ast.ElementType.ELSE
 import com.pinterest.ktlint.core.ast.ElementType.ELVIS
@@ -33,14 +33,12 @@ import com.pinterest.ktlint.core.ast.ElementType.KDOC
 import com.pinterest.ktlint.core.ast.ElementType.KDOC_END
 import com.pinterest.ktlint.core.ast.ElementType.KDOC_LEADING_ASTERISK
 import com.pinterest.ktlint.core.ast.ElementType.KDOC_START
-import com.pinterest.ktlint.core.ast.ElementType.LAMBDA_EXPRESSION
 import com.pinterest.ktlint.core.ast.ElementType.LBRACE
 import com.pinterest.ktlint.core.ast.ElementType.LBRACKET
 import com.pinterest.ktlint.core.ast.ElementType.LITERAL_STRING_TEMPLATE_ENTRY
 import com.pinterest.ktlint.core.ast.ElementType.LONG_STRING_TEMPLATE_ENTRY
 import com.pinterest.ktlint.core.ast.ElementType.LPAR
 import com.pinterest.ktlint.core.ast.ElementType.LT
-import com.pinterest.ktlint.core.ast.ElementType.OBJECT_LITERAL
 import com.pinterest.ktlint.core.ast.ElementType.OPEN_QUOTE
 import com.pinterest.ktlint.core.ast.ElementType.OPERATION_REFERENCE
 import com.pinterest.ktlint.core.ast.ElementType.PARENTHESIZED
@@ -54,7 +52,6 @@ import com.pinterest.ktlint.core.ast.ElementType.SECONDARY_CONSTRUCTOR
 import com.pinterest.ktlint.core.ast.ElementType.SHORT_STRING_TEMPLATE_ENTRY
 import com.pinterest.ktlint.core.ast.ElementType.STRING_TEMPLATE
 import com.pinterest.ktlint.core.ast.ElementType.SUPER_TYPE_CALL_ENTRY
-import com.pinterest.ktlint.core.ast.ElementType.SUPER_TYPE_ENTRY
 import com.pinterest.ktlint.core.ast.ElementType.SUPER_TYPE_LIST
 import com.pinterest.ktlint.core.ast.ElementType.THEN
 import com.pinterest.ktlint.core.ast.ElementType.TYPE_ARGUMENT_LIST
@@ -62,8 +59,6 @@ import com.pinterest.ktlint.core.ast.ElementType.TYPE_CONSTRAINT_LIST
 import com.pinterest.ktlint.core.ast.ElementType.TYPE_PARAMETER_LIST
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_ARGUMENT
 import com.pinterest.ktlint.core.ast.ElementType.VALUE_ARGUMENT_LIST
-import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER
-import com.pinterest.ktlint.core.ast.ElementType.VALUE_PARAMETER_LIST
 import com.pinterest.ktlint.core.ast.ElementType.WHEN_ENTRY
 import com.pinterest.ktlint.core.ast.ElementType.WHERE_KEYWORD
 import com.pinterest.ktlint.core.ast.ElementType.WHITE_SPACE
@@ -73,7 +68,6 @@ import com.pinterest.ktlint.core.ast.isPartOfComment
 import com.pinterest.ktlint.core.ast.isWhiteSpace
 import com.pinterest.ktlint.core.ast.isWhiteSpaceWithNewline
 import com.pinterest.ktlint.core.ast.isWhiteSpaceWithoutNewline
-import com.pinterest.ktlint.core.ast.nextCodeLeaf
 import com.pinterest.ktlint.core.ast.nextCodeSibling
 import com.pinterest.ktlint.core.ast.nextLeaf
 import com.pinterest.ktlint.core.ast.nextSibling
@@ -81,9 +75,6 @@ import com.pinterest.ktlint.core.ast.parent
 import com.pinterest.ktlint.core.ast.prevCodeLeaf
 import com.pinterest.ktlint.core.ast.prevCodeSibling
 import com.pinterest.ktlint.core.ast.prevLeaf
-import com.pinterest.ktlint.core.ast.prevSibling
-import com.pinterest.ktlint.core.ast.upsertWhitespaceAfterMe
-import com.pinterest.ktlint.core.ast.upsertWhitespaceBeforeMe
 import com.pinterest.ktlint.core.ast.visit
 import com.pinterest.ktlint.core.initKtLintKLogger
 import com.pinterest.ktlint.ruleset.standard.IndentationRule.IndentContext.Block
@@ -98,28 +89,31 @@ import org.jetbrains.kotlin.com.intellij.psi.impl.source.tree.LeafPsiElement
 import org.jetbrains.kotlin.com.intellij.psi.tree.IElementType
 import org.jetbrains.kotlin.com.intellij.psi.tree.TokenSet
 import org.jetbrains.kotlin.psi.KtStringTemplateExpression
-import org.jetbrains.kotlin.psi.KtSuperTypeList
 import org.jetbrains.kotlin.psi.psiUtil.leaves
 
 private val logger = KotlinLogging.logger {}.initKtLintKLogger()
 
 /**
- * ktlint's rule that checks & corrects indentation.
- *
- * To keep things simple, we walk the AST twice:
- * - 1st pass - insert missing newlines (e.g. between parentheses of a multi-line function call)
- * - 2st pass - correct indentation
+ * Checks & correct indentation
  *
  * Current limitations:
  * - "all or nothing" (currently, rule can only be disabled for an entire file)
  */
-class IndentationRule : Rule(
-    id = "indent",
-    visitorModifiers = setOf(
-        VisitorModifier.RunOnRootNodeOnly,
-        VisitorModifier.RunAsLateAsPossible
-    )
-) {
+@OptIn(FeatureInAlphaState::class)
+public class IndentationRule :
+    Rule(
+        id = "indent",
+        visitorModifiers = setOf(
+            VisitorModifier.RunOnRootNodeOnly,
+            VisitorModifier.RunAsLateAsPossible
+        )
+    ),
+    UsesEditorConfigProperties {
+    override val editorConfigProperties: List<UsesEditorConfigProperties.EditorConfigProperty<*>> =
+        listOf(
+            indentSizeProperty,
+            indentStyleProperty
+        )
 
     private companion object {
         private val lTokenSet = TokenSet.create(LPAR, LBRACE, LBRACKET, LT)
@@ -138,7 +132,6 @@ class IndentationRule : Rule(
         expectedIndent = 0
     }
 
-    private var editorConfig = EditorConfig.UNINITIALIZED
     private var indentConfig = IndentConfig.DEFAULT_INDENT_CONFIG
 
     override fun visit(
@@ -146,345 +139,20 @@ class IndentationRule : Rule(
         autoCorrect: Boolean,
         emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
     ) {
-        editorConfig = node.loadEditorConfig()
-        indentConfig = editorConfig.loadIndentConfig()
+        indentConfig = IndentConfig(
+            indentStyle = node.getEditorConfigValue(indentStyleProperty),
+            tabWidth = node.getEditorConfigValue(indentSizeProperty)
+        )
         if (indentConfig.disabled) {
             return
         }
-        reset()
-        logger.trace { "phase: rearrangement (auto correction ${if (autoCorrect) "on" else "off"})" }
-        // step 1: insert newlines (if/where needed)
-        var emitted = false
-        rearrange(node, autoCorrect) { offset, errorMessage, canBeAutoCorrected ->
-            emitted = true
-            emit(offset, errorMessage, canBeAutoCorrected)
-        }
-        if (emitted && autoCorrect) {
-            logger.trace {
-                "indenting:\n" +
-                    node
-                        .text
-                        .split("\n")
-                        .mapIndexed { i, v -> "\t${i + 1}: $v" }
-                        .joinToString("\n")
-            }
-        }
-        reset()
-        logger.trace { "phase: indentation" }
-        // step 2: correct indentation
-        indent(node, autoCorrect, emit)
 
+        reset()
+        indent(node, autoCorrect, emit)
         // The expectedIndent should never be negative. If so, it is very likely that ktlint crashes at runtime when
         // autocorrecting is executed while no error occurs with linting only. Such errors often are not found in unit
         // tests, as the examples are way more simple than realistic code.
         assert(expectedIndent >= 0)
-    }
-
-    private fun rearrange(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        node.visit { n ->
-            when (n.elementType) {
-                LPAR, LBRACE, LBRACKET -> rearrangeBlock(n, autoCorrect, emit) // TODO: LT
-                SUPER_TYPE_LIST -> rearrangeSuperTypeList(n, autoCorrect, emit)
-                VALUE_PARAMETER_LIST, VALUE_ARGUMENT_LIST -> rearrangeValueList(n, autoCorrect, emit)
-                ARROW -> rearrangeArrow(n, autoCorrect, emit)
-                WHITE_SPACE -> line += n.text.count { it == '\n' }
-                CLOSING_QUOTE -> rearrangeClosingQuote(n, autoCorrect, emit)
-            }
-        }
-    }
-
-    private fun rearrangeBlock(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        val rElementType = matchingRToken[node.elementType]
-        var newlineInBetween = false
-        var parameterListInBetween = false
-        var numberOfArgs = 0
-        var firstArg: ASTNode? = null
-        // matching ), ] or }
-        val r = node.nextSibling {
-            val isValueArgument = it.elementType == VALUE_ARGUMENT
-            val hasLineBreak = if (isValueArgument) it.hasLineBreak(LAMBDA_EXPRESSION, FUN) else it.hasLineBreak()
-            newlineInBetween = newlineInBetween || hasLineBreak
-            parameterListInBetween = parameterListInBetween || it.elementType == VALUE_PARAMETER_LIST
-            if (isValueArgument) {
-                numberOfArgs++
-                firstArg = it
-            }
-            it.elementType == rElementType
-        }!!
-        if (
-            !newlineInBetween ||
-            // keep { p ->
-            // }
-            (node.elementType == LBRACE && parameterListInBetween) ||
-            // keep ({
-            // }) and (object : C {
-            // })
-            (
-                numberOfArgs == 1 &&
-                    firstArg?.firstChildNode?.elementType
-                    ?.let { it == OBJECT_LITERAL || it == LAMBDA_EXPRESSION } == true
-                )
-        ) {
-            return
-        }
-        if (!node.nextCodeLeaf()?.prevLeaf {
-            // Skip comments, whitespace, and empty nodes
-            !it.isPartOfComment() &&
-                !it.isWhiteSpaceWithoutNewline() &&
-                it.textLength > 0
-        }.isWhiteSpaceWithNewline() &&
-            // IDEA quirk:
-            // if (true &&
-            //     true
-            // ) {
-            // }
-            // instead of
-            // if (
-            //     true &&
-            //     true
-            // ) {
-            // }
-            node.treeNext?.elementType != CONDITION
-        ) {
-            requireNewlineAfterLeaf(node, autoCorrect, emit)
-        }
-        if (!r.prevLeaf().isWhiteSpaceWithNewline()) {
-            requireNewlineBeforeLeaf(r, autoCorrect, emit)
-        }
-    }
-
-    private fun rearrangeSuperTypeList(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        val entries = (node.psi as KtSuperTypeList).entries
-        if (
-            node.textContains('\n') &&
-            entries.size > 1 &&
-            // e.g.
-            //
-            // class A : B, C,
-            //     D
-            // or
-            // class A : B, C({
-            // }), D
-            //
-            // but not
-            //
-            // class A : B, C, D({
-            // })
-            !(
-                entries.dropLast(1).all { it.elementType == SUPER_TYPE_ENTRY } &&
-                    entries.last().elementType == SUPER_TYPE_CALL_ENTRY
-                )
-        ) {
-            // put space after :
-            if (!node.prevLeaf().isWhiteSpaceWithNewline()) {
-                val colon = node.prevCodeLeaf()!!
-                if (
-                    !colon.prevLeaf().isWhiteSpaceWithNewline() &&
-                    colon.prevCodeLeaf().let { it?.elementType != RPAR || !it.prevLeaf().isWhiteSpaceWithNewline() }
-                ) {
-                    requireNewlineAfterLeaf(colon, autoCorrect, emit)
-                }
-            }
-            // put entries on separate lines
-            // TODO: group emit()s below with the one above into one (similar to ParameterListWrappingRule)
-            for (c in node.children()) {
-                if (c.elementType == COMMA && !c.treeNext.isWhiteSpaceWithNewline()) {
-                    requireNewlineAfterLeaf(c, autoCorrect, emit)
-                }
-            }
-        }
-    }
-
-    private fun rearrangeValueList(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        for (c in node.children()) {
-            val hasLineBreak = when (c.elementType) {
-                VALUE_ARGUMENT -> c.hasLineBreak(LAMBDA_EXPRESSION, FUN)
-                VALUE_PARAMETER, ANNOTATION -> c.hasLineBreak()
-                else -> false
-            }
-            if (hasLineBreak) {
-                // rearrange
-                //
-                // a, b, value(
-                // ), c, d
-                //
-                // to
-                //
-                // a, b,
-                // value(
-                // ),
-                // c, d
-
-                // insert \n in front of multi-line value
-                val prevSibling = c.prevSibling { it.elementType != WHITE_SPACE }
-                if (
-                    prevSibling?.elementType == COMMA &&
-                    !prevSibling.treeNext.isWhiteSpaceWithNewline()
-                ) {
-                    requireNewlineAfterLeaf(prevSibling, autoCorrect, emit)
-                }
-                // insert \n after multi-line value
-                val nextSibling = c.nextSibling { it.elementType != WHITE_SPACE }
-                if (
-                    nextSibling?.elementType == COMMA &&
-                    !nextSibling.treeNext.isWhiteSpaceWithNewline() &&
-                    // value(
-                    // ), // a comment
-                    // c, d
-                    nextSibling.treeNext?.treeNext?.psi !is PsiComment
-                ) {
-                    requireNewlineAfterLeaf(nextSibling, autoCorrect, emit)
-                }
-            }
-        }
-    }
-
-    private fun rearrangeClosingQuote(
-        n: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        val treeParent = n.treeParent
-        if (treeParent.elementType == STRING_TEMPLATE) {
-            val treeParentPsi = treeParent.psi as KtStringTemplateExpression
-            if (treeParentPsi.isMultiLine() && n.treePrev.text.isNotBlank()) {
-                // rewriting
-                // """
-                //     text
-                // _""".trimIndent()
-                // to
-                // """
-                //     text
-                // _
-                // """.trimIndent()
-                emit(
-                    n.startOffset,
-                    "Missing newline before \"\"\"",
-                    true
-                )
-                if (autoCorrect) {
-                    n as LeafPsiElement
-                    n.rawInsertBeforeMe(LeafPsiElement(REGULAR_STRING_PART, "\n"))
-                }
-                logger.trace { "$line: " + (if (!autoCorrect) "would have " else "") + "inserted newline before (closing) \"\"\"" }
-            }
-        }
-    }
-
-    private fun mustBeFollowedByNewline(node: ASTNode): Boolean {
-        // find EOL token (last token before \n)
-        // if token is in lTokenSet
-        //     find matching rToken
-        //     return true if there is no newline after the rToken
-        // return false
-        val p = node.treeParent
-        val nextCodeSibling = node.nextCodeSibling() // e.g. BINARY_EXPRESSION
-        var lToken = nextCodeSibling?.nextLeaf { it.isWhiteSpaceWithNewline() }?.prevCodeLeaf()
-        if (lToken != null && lToken.elementType !in lTokenSet) {
-            // special cases:
-            // x = y.f({ z ->
-            // })
-            // x = y.f(0, 1,
-            // 2, 3)
-            lToken = lToken.prevLeaf { it.elementType in lTokenSet || it == node }
-        }
-        if (lToken != null && lToken.elementType in lTokenSet) {
-            val rElementType = matchingRToken[lToken.elementType]
-            val rToken = lToken.nextSibling { it.elementType == rElementType }
-            return rToken?.treeParent == lToken.treeParent
-        }
-        if (nextCodeSibling?.textContains('\n') == false) {
-            return true
-        }
-        return false
-    }
-
-    private fun rearrangeArrow(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        val p = node.treeParent
-        if (
-            // check
-            // `{ p -> ... }`
-            // and
-            // `when { m -> ... }`
-            // only
-            p.elementType.let { it != FUNCTION_LITERAL && it != WHEN_ENTRY } ||
-            // ... and only if expression after -> spans multiple lines
-            !p.textContains('\n') ||
-            // permit
-            // when {
-            //     m -> 0 + d({
-            //     })
-            // }
-            (p.elementType == WHEN_ENTRY && mustBeFollowedByNewline(node)) ||
-            // permit
-            // when (this) {
-            //     in 0x1F600..0x1F64F, // Emoticons
-            //     0x200D // Zero-width Joiner
-            //     -> true
-            // }
-            (p.elementType == WHEN_ENTRY && node.prevLeaf()?.textContains('\n') == true)
-        ) {
-            return
-        }
-        if (!node.nextCodeLeaf()?.prevLeaf().isWhiteSpaceWithNewline()) {
-            requireNewlineAfterLeaf(node, autoCorrect, emit)
-        }
-        val r = node.nextSibling { it.elementType == RBRACE } ?: return
-        if (!r.prevLeaf().isWhiteSpaceWithNewline()) {
-            requireNewlineBeforeLeaf(r, autoCorrect, emit)
-        }
-    }
-
-    private fun requireNewlineBeforeLeaf(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        emit(
-            node.startOffset - 1,
-            """Missing newline before "${node.text}"""",
-            true
-        )
-        logger.trace { "$line: " + ((if (!autoCorrect) "would have " else "") + "inserted newline before ${node.text}") }
-        if (autoCorrect) {
-            (node.psi as LeafPsiElement).upsertWhitespaceBeforeMe("\n ")
-        }
-    }
-
-    private fun requireNewlineAfterLeaf(
-        node: ASTNode,
-        autoCorrect: Boolean,
-        emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
-    ) {
-        emit(
-            node.startOffset + 1,
-            """Missing newline after "${node.text}"""",
-            true
-        )
-        logger.trace { "$line: " + (if (!autoCorrect) "would have " else "") + "inserted newline after ${node.text}" }
-        if (autoCorrect) {
-            (node.psi as LeafPsiElement).upsertWhitespaceAfterMe("\n ")
-        }
     }
 
     private class IndentContext {
@@ -931,98 +599,102 @@ class IndentationRule : Rule(
         autoCorrect: Boolean,
         emit: (offset: Int, errorMessage: String, canBeAutoCorrected: Boolean) -> Unit
     ) {
-        val psi = node.psi as KtStringTemplateExpression
-        if (psi.isMultiLine()) {
-            if (node.containsMixedIndentationCharacters()) {
-                // It can not be determined with certainty how mixed indentation characters should be interpreted.
-                // The trimIndent function handles tabs and spaces equally (one tabs equals one space) while the user
-                // might expect that the tab size in the indentation is more than one space.
-                emit(
-                    node.startOffset,
-                    "Indentation of multiline string should not contain both tab(s) and space(s)",
-                    false
-                )
-                return
-            }
-
-            val prefixLength = node.children()
-                .filterNot { it.elementType == OPEN_QUOTE }
-                .filterNot { it.elementType == CLOSING_QUOTE }
-                .filter { it.prevLeaf()?.text == "\n" }
-                .filterNot { it.text == "\n" }
-                .let { indents ->
-                    val indentsExceptBlankIndentBeforeClosingQuote = indents
-                        .filterNot { it.isIndentBeforeClosingQuote() }
-                    if (indentsExceptBlankIndentBeforeClosingQuote.count() > 0) {
-                        indentsExceptBlankIndentBeforeClosingQuote
-                    } else {
-                        indents
-                    }
+        node
+            .let { it.psi as KtStringTemplateExpression }
+            .takeIf { it.isFollowedByTrimIndent() || it.isFollowedByTrimMargin() }
+            ?.takeIf { it.isMultiLine() }
+            ?.let {
+                if (node.containsMixedIndentationCharacters()) {
+                    // It can not be determined with certainty how mixed indentation characters should be interpreted.
+                    // The trimIndent function handles tabs and spaces equally (one tabs equals one space) while the user
+                    // might expect that the tab size in the indentation is more than one space.
+                    emit(
+                        node.startOffset,
+                        "Indentation of multiline string should not contain both tab(s) and space(s)",
+                        false
+                    )
+                    return
                 }
-                .map { it.text.indentLength() }
-                .minOrNull() ?: 0
 
-            val correctedExpectedIndent = if (node.prevLeaf()?.text == "\n") {
-                // In case the opening quotes are placed at the start of the line, then expect all lines inside the
-                // string literal and the closing quotes to have no indent as well.
-                0
-            } else {
-                expectedIndent
-            }
-            val expectedIndentation = indentConfig.indent.repeat(correctedExpectedIndent)
-            val expectedPrefixLength = correctedExpectedIndent * indentConfig.indent.length
-            node.children()
-                .forEach {
-                    if (it.prevLeaf()?.text == "\n" &&
-                        (
-                            it.isLiteralStringTemplateEntry() ||
-                                it.isVariableStringTemplateEntry() ||
-                                it.isClosingQuote()
-                            )
-                    ) {
-                        val (actualIndent, actualContent) =
-                            if (it.isIndentBeforeClosingQuote()) {
-                                it.text.splitIndentAt(it.text.length)
-                            } else if (it.isVariableStringTemplateEntry() && it.isFirstNonBlankElementOnLine()) {
-                                it.getFirstElementOnSameLine().text.splitIndentAt(expectedPrefixLength)
-                            } else {
-                                it.text.splitIndentAt(prefixLength)
-                            }
-                        if (indentConfig.containsUnexpectedIndentChar(actualIndent)) {
-                            val offsetFirstWrongIndentChar = indentConfig.indexOfFirstUnexpectedIndentChar(actualIndent)
-                            emit(
-                                it.startOffset + offsetFirstWrongIndentChar,
-                                "Unexpected '${indentConfig.unexpectedIndentCharDescription}' character(s) in margin of multiline string",
-                                true
-                            )
-                            if (autoCorrect) {
-                                (it.firstChildNode as LeafPsiElement).rawReplaceWithText(
-                                    expectedIndentation + actualContent
+                val prefixLength = node.children()
+                    .filterNot { it.elementType == OPEN_QUOTE }
+                    .filterNot { it.elementType == CLOSING_QUOTE }
+                    .filter { it.prevLeaf()?.text == "\n" }
+                    .filterNot { it.text == "\n" }
+                    .let { indents ->
+                        val indentsExceptBlankIndentBeforeClosingQuote = indents
+                            .filterNot { it.isIndentBeforeClosingQuote() }
+                        if (indentsExceptBlankIndentBeforeClosingQuote.count() > 0) {
+                            indentsExceptBlankIndentBeforeClosingQuote
+                        } else {
+                            indents
+                        }
+                    }
+                    .map { it.text.indentLength() }
+                    .minOrNull() ?: 0
+
+                val correctedExpectedIndent = if (node.prevLeaf()?.text == "\n") {
+                    // In case the opening quotes are placed at the start of the line, then expect all lines inside the
+                    // string literal and the closing quotes to have no indent as well.
+                    0
+                } else {
+                    expectedIndent
+                }
+                val expectedIndentation = indentConfig.indent.repeat(correctedExpectedIndent)
+                val expectedPrefixLength = correctedExpectedIndent * indentConfig.indent.length
+                node.children()
+                    .forEach {
+                        if (it.prevLeaf()?.text == "\n" &&
+                            (
+                                it.isLiteralStringTemplateEntry() ||
+                                    it.isVariableStringTemplateEntry() ||
+                                    it.isClosingQuote()
                                 )
-                            }
-                        } else if (actualIndent != expectedIndentation && it.isIndentBeforeClosingQuote()) {
-                            // It is a deliberate choice not to fix the indents inside the string literal except the line which only contains
-                            // the closing quotes.
-                            emit(
-                                it.startOffset,
-                                "Unexpected indent of multiline string closing quotes",
-                                true
-                            )
-                            if (autoCorrect) {
-                                if (it.firstChildNode == null) {
-                                    (it as LeafPsiElement).rawInsertBeforeMe(
-                                        LeafPsiElement(REGULAR_STRING_PART, expectedIndentation)
-                                    )
+                        ) {
+                            val (actualIndent, actualContent) =
+                                if (it.isIndentBeforeClosingQuote()) {
+                                    it.text.splitIndentAt(it.text.length)
+                                } else if (it.isVariableStringTemplateEntry() && it.isFirstNonBlankElementOnLine()) {
+                                    it.getFirstElementOnSameLine().text.splitIndentAt(expectedPrefixLength)
                                 } else {
+                                    it.text.splitIndentAt(prefixLength)
+                                }
+                            if (indentConfig.containsUnexpectedIndentChar(actualIndent)) {
+                                val offsetFirstWrongIndentChar =
+                                    indentConfig.indexOfFirstUnexpectedIndentChar(actualIndent)
+                                emit(
+                                    it.startOffset + offsetFirstWrongIndentChar,
+                                    "Unexpected '${indentConfig.unexpectedIndentCharDescription}' character(s) in margin of multiline string",
+                                    true
+                                )
+                                if (autoCorrect) {
                                     (it.firstChildNode as LeafPsiElement).rawReplaceWithText(
                                         expectedIndentation + actualContent
                                     )
                                 }
+                            } else if (actualIndent != expectedIndentation && it.isIndentBeforeClosingQuote()) {
+                                // It is a deliberate choice not to fix the indents inside the string literal except the line which only contains
+                                // the closing quotes.
+                                emit(
+                                    it.startOffset,
+                                    "Unexpected indent of multiline string closing quotes",
+                                    true
+                                )
+                                if (autoCorrect) {
+                                    if (it.firstChildNode == null) {
+                                        (it as LeafPsiElement).rawInsertBeforeMe(
+                                            LeafPsiElement(REGULAR_STRING_PART, expectedIndentation)
+                                        )
+                                    } else {
+                                        (it.firstChildNode as LeafPsiElement).rawReplaceWithText(
+                                            expectedIndentation + actualContent
+                                        )
+                                    }
+                                }
                             }
                         }
                     }
-                }
-        }
+            }
     }
 
     private fun KtStringTemplateExpression.isMultiLine(): Boolean {
@@ -1256,16 +928,6 @@ class IndentationRule : Rule(
         return false
     }
 
-    private fun ASTNode.hasLineBreak(vararg ignoreElementTypes: IElementType): Boolean {
-        if (isWhiteSpaceWithNewline()) return true
-        return if (ignoreElementTypes.isEmpty()) {
-            textContains('\n')
-        } else {
-            elementType !in ignoreElementTypes &&
-                children().any { c -> c.textContains('\n') && c.elementType !in ignoreElementTypes }
-        }
-    }
-
     private fun ASTNode.containsMixedIndentationCharacters(): Boolean {
         assert((this.psi as KtStringTemplateExpression).isMultiLine())
         val nonBlankLines = this
@@ -1362,3 +1024,11 @@ private fun String.splitIndentAt(index: Int): Pair<String, String> {
         second = this.substring(safeIndex)
     )
 }
+
+private fun KtStringTemplateExpression.isFollowedByTrimIndent() = isFollowedBy("trimIndent()")
+
+private fun KtStringTemplateExpression.isFollowedByTrimMargin() = isFollowedBy("trimMargin()")
+
+private fun KtStringTemplateExpression.isFollowedBy(callExpressionName: String) =
+    this.node.nextSibling { it.elementType != DOT }
+        .let { it?.elementType == CALL_EXPRESSION && it.text == callExpressionName }