+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/NoTrailingSpacesRuleTest.kt
@@ -193,4 +193,52 @@ class NoTrailingSpacesRuleTest {
             )
         )
     }
+
+    @Test
+    fun `Issue 1376 - trailing spaces should not delete blank line inside kdoc`() {
+        val code =
+            """
+            /**
+             Paragraph 1 which should be followed by a blank line.
+
+
+             Paragraph 2 which should have a blank line before it.
+             */
+            class MyClass
+            """.trimIndent()
+        assertThat(NoTrailingSpacesRule().lint(code)).isEmpty()
+        assertThat(NoTrailingSpacesRule().format(code)).isEqualTo(code)
+    }
+
+    @Test
+    fun `Issue 1376 - trailing spaces should be removed from blank line inside kdoc`() {
+        val code =
+            """
+            /**
+             Paragraph 1 which should be followed by a blank line.
+             ${"    "}
+             ${"    "}
+             Paragraph 2 which should have a blank line before it.
+             */
+            class MyClass
+            """.trimIndent()
+        val formattedCode =
+            """
+            /**
+             Paragraph 1 which should be followed by a blank line.
+
+
+             Paragraph 2 which should have a blank line before it.
+             */
+            class MyClass
+            """.trimIndent()
+        assertThat(
+            NoTrailingSpacesRule().format(code)
+        ).isEqualTo(formattedCode)
+        assertThat(
+            NoTrailingSpacesRule().lint(code)
+        ).containsExactly(
+            LintError(3, 1, "no-trailing-spaces", "Trailing space(s)")
+        )
+    }
 }